backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pyjwt==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyPDF2==3.0.1
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
openai==1.10.0
groq==0.4.2
tiktoken==0.5.2
numpy==1.26.3
